+++ lib/tools/hwnd.dart
@@ -1,6 +1,7 @@
 import 'dart:ffi';
 
 import 'package:ffi/ffi.dart';
+import 'package:win/win32_add.dart';
 import 'package:win32/win32.dart';
 
 import 'primitives.dart';
@@ -20,6 +21,50 @@ abstract class Hwnd {
   factory Hwnd.fomHandle(int handle) {
     return _Hwnd(handle);
   }
+
+  static Hwnd? fromProcessID(int processId, {bool wait = false}) {
+    final data = calloc<EnumWindowsData>()
+      ..ref.hwnd = 0
+      ..ref.process_id = processId;
+
+    final callback = Pointer.fromFunction<EnumWindowsProc>(
+      _enumWindowsCallback,
+      0,
+    );
+
+    try {
+       do {
+        EnumWindows(callback, data.address);
+      } while (wait && data.ref.hwnd == 0);
+    } finally {
+      free(data);
+    }
+    return data.ref.hwnd == 0 ? null : Hwnd.fomHandle(data.ref.hwnd);
+  }
+
+  static int _enumWindowsCallback(int hwnd, int lParam) {
+    final data = Pointer<EnumWindowsData>.fromAddress(lParam);
+
+    final process_id = calloc<Uint32>();
+    try {
+      GetWindowThreadProcessId(hwnd, process_id);
+
+      if (process_id.value != data.ref.process_id) {
+        return TRUE;
+      }
+    } finally {
+      free(process_id);
+    }
+
+    final own = GetWindow(hwnd, GW_OWNER);
+    if (own != 0) {
+      return TRUE;
+    }
+
+    data.ref.hwnd = hwnd;
+    return FALSE; // stop enums
+  }
+
 
   void destroy() => DestroyWindow(handle);
 
@@ -62,14 +107,24 @@ abstract class Hwnd {
 
   set rect(Rect newRect) {
     SetWindowPos(
-      handle,
-      NULL,
-      newRect.left,
-      newRect.top,
-      newRect.width,
-      newRect.height,
-      SWP_NOZORDER | SWP_NOACTIVATE,
-    );
+        handle,
+        HWND_TOP,
+
+        newRect.left,
+        newRect.top,
+        newRect.width,
+        newRect.height,
+        SWP_SHOWWINDOW);
+
+    // SetWindowPos(
+    //   handle,
+    //   HWND_TOP,
+    //   newRect.left,
+    //   newRect.top,
+    //   newRect.width,
+    //   newRect.height,
+    //   SWP_NOZORDER | SWP_NOACTIVATE,
+    // );
   }
 
   Rect get clientRect {
@@ -124,6 +179,14 @@ abstract class Hwnd {
     child.parent = this;
     child.rect = Rect(0, 0, thisSize.width, thisSize.height);
     _childContent = child;
-    _childContent?.show();
+    _childContent!.show();
   }
+}
+
+class EnumWindowsData extends Struct {
+  @IntPtr()
+  external int hwnd;
+
+  @IntPtr()
+  external int process_id;
 }
