+++ pubspec.yaml
@@ -11,7 +11,7 @@ environment:
 dependencies:
   ffi: ^1.0.0
   win32: ^2.2.8
-  meta: ^1.7.0
+  meta: ^1.3.0
   path:
 dev_dependencies:
   lints: ^1.0.1
